turned Pubrec Inbound/Outbound Interceptor Handlers into a Duplex Handler
