+++ src/main/java/com/hivemq/bootstrap/netty/ChannelDependencies.java
@@ -162,7 +162,7 @@ public class ChannelDependencies {
     private final ConnackOutboundInterceptorHandler connackOutboundInterceptorHandler;
 
     @NotNull
-    private final PubrecOutboundInterceptorHandler pubrecOutboundInterceptorHandler;
+    private final PubrecInterceptorHandler pubrecInterceptorHandler;
 
     @Inject
     public ChannelDependencies(
@@ -201,7 +201,7 @@ public class ChannelDependencies {
             @NotNull final PublishOutboundInterceptorHandler publishOutboundInterceptorHandler,
             @NotNull final ConnectInboundInterceptorHandler connectInboundInterceptorHandler,
             @NotNull final ConnackOutboundInterceptorHandler connackOutboundInterceptorHandler,
-            @NotNull final PubrecOutboundInterceptorHandler pubrecOutboundInterceptorHandler) {
+            @NotNull final PubrecInterceptorHandler pubrecInterceptorHandler) {
 
 
         this.statisticsInitializer = statisticsInitializer;
@@ -239,7 +239,7 @@ public class ChannelDependencies {
         this.publishOutboundInterceptorHandler = publishOutboundInterceptorHandler;
         this.connectInboundInterceptorHandler = connectInboundInterceptorHandler;
         this.connackOutboundInterceptorHandler = connackOutboundInterceptorHandler;
-        this.pubrecOutboundInterceptorHandler = pubrecOutboundInterceptorHandler;
+        this.pubrecInterceptorHandler = pubrecInterceptorHandler;
     }
 
     @NotNull
@@ -418,7 +418,7 @@ public class ChannelDependencies {
     }
 
     @NotNull
-    public PubrecOutboundInterceptorHandler getPubrecOutboundInterceptorHandler() {
-        return pubrecOutboundInterceptorHandler;
+    public PubrecInterceptorHandler getPubrecOutboundInterceptorHandler() {
+        return pubrecInterceptorHandler;
     }
 }
